fixed: segment size of each file should be int64 (thx jmarshall)
git-svn-id: https://xbmc.svn.sourceforge.net/svnroot/xbmc/branches/linuxport/XBMC@17044 568bbfeb-2a22-0410-94d2-cc84cf5bfa90

+++ xbmc/cores/dvdplayer/DVDInputStreams/DVDInputStreamStack.h
@@ -43,8 +43,8 @@ protected:
 
   struct TSeg
   {
-    TFile file;
-    int   length;
+    TFile   file;
+    __int64 length;
   };
 
   typedef std::vector<TSeg> TSegVec;
